refactor: simplify refresh balance triggers (#6232)
### Description

This PR consolidates the way we are refreshing the token balances. I
think we let this evolve organically for some time and it has become
buggy and inconsistent.

Currently, before this change:
- we have a bunch of watcher functions that seem overcomplicated
(dispatching intermediate actions to trigger other sagas)
- we need to remember to dispatch the `fetchTokenBalances` action after
every flow that initiates a transaction. i think it is probably
unavoidable to need to remember to trigger a balance fetch, but
currently we dispatch a success action for the flow and a separate
action for the balances (the dreaded double dispatch)
- we forget to refresh balances in a bunch of places (e.g. swap flow)
which means users can try (and fail) to execute other transactions with
outdated balances.

This change:
- remove the `fetchTokenBalances` action entirely and instead, trigger
the balance refresh sagas on specific actions. i've tried to document
them all.
- there is a loading state that seems unnecessary in the `home` reducer,
most flows were never allowing this loading state to be set. only the
send flow was, i think that's a bug. balance refreshes should be a
background task and not affect whether we show or hide the balance....
- define balance refresh as a combination of refreshing token balances,
exchange rates (the total balance is wrong if the exchange rate is
outdated), positions so trigger all those sagas

### Test plan

I should see a network request to refresh token balance, exchange rate,
positions, shortcuts when I have:
- completed onboarding
- received a payment (transaction feed updated)
- executed a transaction via any app flow

### Related issues

- Related to RET-1241

### Backwards compatibility

yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)