fix(route): Fix issue with initial route, clean up onboarding feature config (#6218)
### Description

Currently, even if CPV is disabled in onboarding (or disabled everywhere
else in the app as well, as is the case with MS apps), if a user exits
the app during onboarding under certain conditions and then restarts the
app, they'll be sent to the CPV in onboarding screen.

This PR updates the `getInitialRoute` logic to avoid sending the user to
the CPV screen if CPV is disabled. In addition, the onboarding config
parameters regarding CAB setup & restore have been consolidated.
Basically, there are now two config params. The `CloudBackupRestore`
param makes CAB available as a feature throughout the app, meaning in
both onboarding restore flow + in settings, but does _not_ show the
setup in onboarding. If (and only if) `CloudBackupRestore` has been set
to true, `CloudBackupSetupInOnboarding` will show CAB in the "new
wallet" onboarding flow. We decided to set it up this way in order to
prevent scenarios where CAB is enabled in new wallet setup, but not
available as a restore option/in the settings screen.

### Test plan

Unit and manual tested.

### Related issues

- Fixes #[issue number here]

### Backwards compatibility

<!-- Brief explanation of why these changes are/are not backwards
compatible. -->

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)