feat: add percentage amount options to earn and send flows (#6240)
### Description

This PR:
- adds new component that has the percentage amount options to the earn
and send flows. the component uses JS animations as this was the easiest
thing to implement.
- removes the existing "max" button on these flows
- updates analytics and tests

### Test plan

On iOS:


https://github.com/user-attachments/assets/5b79b872-4192-4a7f-90fb-a99e6993c969


On Android


https://github.com/user-attachments/assets/eacaf60e-cbeb-4a1c-b132-7bf7d2f6baf0



### Related issues

- Related to RET-1236

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)