fix(sends): Enforce token override when sending specific token via QR code (#6250)
### Description

For
[ACT-1462](https://linear.app/valora/issue/ACT-1462/send-flow-qr-scanner-not-preserving-token-id-override).
Updates the send flow + QR path to pass through a token override when
entering via a token details screen.

### Test plan

Unit tests updated + manual tests. See video below.



https://github.com/user-attachments/assets/ee7a3aea-93de-4ab9-bb91-cd6fed629d61


### Related issues

- Fixes
[ACT-1462](https://linear.app/valora/issue/ACT-1462/send-flow-qr-scanner-not-preserving-token-id-override).

### Backwards compatibility

<!-- Brief explanation of why these changes are/are not backwards
compatible. -->

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)