feat(TokenEnterAmount): add new flow to EnterAmount.tsx (#6245)
### Description
3/5 PR for new Enter Amount component. This PR uses the new component
and hook in the Send flow. A lot of tests are adjusted to follow the new
flow.

`AmountInput` component was moved purely for compatibility with
`EarnEnterAmount` component. This is an outdated boilerplate code which
will be removed in the next PR (4/5). Most of the line changes is this
specific component and a lot of tests' adjusting. There's really not
many changes other than replacing boilerplate code with the same code
from a reusable hook.

### Test plan
Changes a bunch of test files to relate to the new changes. All the
tests related to `MAX` button are skipped for now.

<video
src="https://github.com/user-attachments/assets/6d3e286e-393a-4061-85e8-6f5d04713dba"
width="500"></video>

### Related issues

- Relates to RET-1208
### Backwards compatibility

Yes
### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)

---------

Co-authored-by: Kathy Luo <kathyluo18@gmail.com>