chore: update copy for deposit to unknown app (#6298)
### Description

It may be unavoidable that some dapps that we do not know about create
transactions that are parsed as deposits / withdraws etc. The "unknown
dapp" message reads a little bit like an error (and was flagged as such
at bugapalooza,
[context](https://valora-app.slack.com/archives/C04B61SJ6DS/p1733248109384319?thread_ts=1733245811.844689&cid=C04B61SJ6DS))
so this PR updates the copy. Open for other suggestions too.

### Test plan

n/a

### Related issues
n/a

### Backwards compatibility

Y
### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)