chore(cleanup): Remove onboarding_terms_and_conditions experiment (#6301)
### Description

Test experiment added to enum/constants so that CI can pass. With those
being empty tests would fail, and looking at git history could not find
a time when it was empty. This seemed like the easiest solution, and
added a comment explaining.

### Test plan

CI

### Related issues

- Part of ACT-1443

### Backwards compatibility

Yes
### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- N/A