fix(earn): Fix Earn Deposit bottom sheet bug on Android (#6236)
### Description

For
[ACT-1398](https://linear.app/valora/issue/ACT-1398/android-pin-modal-display-obscured-earn-review-bottom-sheet).
Fixes an issue where the deposit bottom sheet was obscuring the PIN
entry screen.

### Test plan

Unit and manual tested, see video below:


https://github.com/user-attachments/assets/27140fb2-3ad5-4363-b458-7a4e8a7f1bb1

### Related issues

- Fixes #[issue number here]

### Backwards compatibility

<!-- Brief explanation of why these changes are/are not backwards
compatible. -->

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)

---------

Co-authored-by: Tom McGuire <Mcgtom10@gmail.com>