fix: broken custom payment with Bidali (#6383)
### Description

This fixes the custom payment integration with Bidali that was broken by
https://github.com/valora-inc/wallet/pull/5692
Because it renamed `window.valora` to `window.walletApp`.

Note: it would be nice to make this Bidali payment flow work multichain
too. Right now it's only for the Celo network.

### Test plan

**Before**


https://github.com/user-attachments/assets/f2d0946a-376e-4b2b-b029-d25e5659ef74

It's not showing the custom payment integration. Only WalletConnect or
the QR code (which is not really usable as is in the mobile-to-mobile
flow).
Note: the payment QR code causes an incorrect conversion, but that's
another story, we'll tackle that separately.

**After**


https://github.com/user-attachments/assets/abe723c9-0a33-4737-81ac-de1834ab02b9

It's showing the custom payment integration "Pay with Valora".

### Related issues

- Part of RET-1292

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)