fix: new 2025 year causing one of the PointsHistoryBottomSheet tests to fail (#6392)
### Description
As per the title. This change tweaks the test to not expect an exact
string but to just check for a partial match.
Context:
[thread](https://valora-app.slack.com/archives/C029Z1QMD7B/p1735811485840799)
GH run with an error:
[run](https://github.com/valora-inc/wallet/actions/runs/12580445142/job/35062355717?pr=6247)

### Test plan
Tests pass

### Related issues
N/A

### Backwards compatibility
Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)