chore(swap): track `swapId` for same-chain swap events (#6398)
### Description

It is expected that each submitted same-chain swap should result either
in success or an error.

From an analytics perspective, the number of `swap_review_submit` events
must equal the sum of `swap_execute_success` and `swap_execute_error`
events.

Unfortunately, currently we have more `swap_reveiw_submit` events than
the sum of success and error ones. It suggests that either we send
duplicated submit events, or miss some success or error events.

This PR adds a `swapId` property shared between sequential swap events
to ensure we can trace them in analytics.

Context:
https://valora-app.slack.com/archives/C029Z1QMD7B/p1733859318155609

### Test plan

Updated unit tests

### Related issues

- Related to RET-1276

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)