feat: monorepo setup with basic workflows working (#3)
### Description

Monorepo setup with basic workflows:
- test
- check

And skipped/disabled other workflows that are currently unnecessary:
- crowdin and translations related
- nightly builds
- license update

The e2e tests will be restored once we add the app to use the runtime in
this repo.

Linting, formatting and knip should now only be run from the root
folder.

### Test plan

CI passes

### Related issues

- Part of RET-1279

### Backwards compatibility

N/A

### Network scalability

N/A