chore(swap): track swap cancel (#6399)
### Description

When the swap is cancelled by the user we [aren't tracking
anything](https://github.com/valora-inc/wallet/blob/e72427d4ecad0f011fa859067adb3e78c3884d3f/src/swap/saga.ts#L297-L301).
This fixes it.

After this fix, we can expect that the number of submitted same-chain
swaps should equal to sum of successful, errored, _and cancelled_ swaps.

A new event `swap_cancel` is introduced instead of reusing
`swap_execute_error` to not contribute to the swap error count.

### Test plan

Updated unit test

### Related issues

- Related to RET-1276

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)