fix: swap amount in fiat is calculated wrong (#6406)
### Description
This PR fixes SwapScreenV2 incorrect fiat calculations when entering
amount in fiat.

P.S. This part will be covered with tests in the follow-up PR for the
sake of not withholding the ability to patch the latest release as it
appears to be more time-consuming.

### Test plan

https://github.com/user-attachments/assets/47e0c263-00c4-4ad3-b529-48839e28f70c

### Related issues
Fixes RET-1296

### Backwards compatibility
Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)