feat: support running without a STATSIG_API_KEY (#6403)
### Description

As the title says.

Before this, running the app without a STATSIG_API_KEY was throwing a
bunch of errors. And crashing in dev mode in the settings.
See
https://github.com/mobilestack-xyz/mobilestack-beefy/pull/2#issuecomment-2573675533

### Test plan

- Tests pass

### Related issues

- Part of RET-1284

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)