feat: use runtime in the example app (#6)
### Description

This updates the example app to use the runtime.

Similar to what was done in the prototype:
https://github.com/mobilestack-xyz/app-expo
Except I've left out firebase for now.

The app now builds and runs on iOS and Android.
Though there's a crash in the setting due to Statsig not being
configured. This is being addressed in
https://github.com/valora-inc/wallet/pull/6403.

Next: make e2e tests work with the example app.

Part of RET-1279