ci: restore iOS e2e test (#8)
### Description

The iOS e2e test now works for the example app.

One thing I noticed is building the iOS app for detox takes longer: it's
now **~7 mins** and previously it was around **~3:30 mins**
([example](https://github.com/valora-inc/wallet/actions/runs/12764358461/job/35576389738)).

I'm not sure if it's because there are a few more native modules to
build with expo or if it's something else.
For instance because the project is re-generated during the workflow,
maybe we don't get as much caching from Xcode compared to what we had
previously.

We could try to enable ccache though to see if it'd help: 

https://docs.expo.dev/versions/latest/sdk/build-properties/#:~:text=Description-,ccacheEnabled,-(optional)

It could also be explained by the new RN version used to build.

Or a combination of all of these.

Part of RET-1279