refactor(colors): deprecate literal colors, add semantic colors to replace 'black' (#6400)
### Description

This PR removes `Colors.black` and replaces the color with semantic
variables that should make it easier to customise colors in the app.
1. add `Colors.textPrimary` to `typeScale` and remove any component
styles that are styling text to black. Unfortunately this adds a lot of
updates to the snapshot tests :(
2. use `Colors.textPrimary` as default icon color (since often icons are
displayed alongside text, it's hard to imagine them needing to have
different defaults)
3. add top and bottom navigation colors to any styles that were set to
black for navigation icons / text
4. use `Colors.backgroundInverse` for styles related to background
colors (and borders that were intended to blend in with background, e.g.
filter chips)
5. use `Colors.shadow` for the one place that declares shadows
6. rename "primary" to "buttonPrimary" since this is what it is used
for.

I did a working session with Kayla to finalise these changes ✅

### Test plan

There should be no user facing changes

### Related issues

- Related to RET-1293

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)