fix: number is formatted wrong when a dot is a grouping separator (#6428)
### Description
When phone number format settings are set to "1.234.567,89" – the number
formatting is all over the place. The issue was an
`unformatNumberForProcessing` function which was only expecting numbers
that were formatted according to the phone's number format settings
while SwapScreenV2 was using numbers sent by a prepare transaction quote
which are formatted as a regular JS number ("1234567.89").

This PR fixes this issue.

### Test plan
<img
src="https://github.com/user-attachments/assets/e7eab81f-0df0-41d9-95f4-64f7f2076f0d"
width="300" />


### Related issues
N/A

### Backwards compatibility
Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)