refactor(colors): refine and rename colors for consistency (#6425)
### Description

The purpose of this PR is to refine the colors enum and choose names
that we are happy with / that are consistent. Since the changes are
applied throughout the project, the size of the change is big but they
are all renames (except for the Button component, where I add the
`buttonTertiary` color to the tertiary button).

### Test plan

n/a no visual changes

### Related issues

- Related to RET-1293

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)