fix(colors): explicitly define colors in places that was missing them (#6426)
### Description

When playing around with changing the colors to another app (Beefy) I
noticed there were components that default to black / white that don't
work well for a darker theme. This PR fixes those places to make sure
the correct color is applied.

### Test plan

n/a
### Related issues

- Related to RET-1293

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)