feat(ReviewTransaction): add initial structure for ReviewTransaction component (#6433)
### Description
This PR adds an initial structure for the new ReviewTransaction
component that is supposed to be used as a common component for all the
review screens. The implementation approach was inspired by the way
components are utilizing composition in libraries like
[shadcn/ui](https://ui.shadcn.com/docs/components/card) and
[MUI](https://mui.com/material-ui/guides/composition/).

[Link to the
design](https://www.figma.com/design/z88kTrIfLyddOCSiAp2dZV/Updates-for-Consistency?node-id=2487-40727&t=2uUaB4bMg2tpNaz0-4)

### Test plan
Added tests to test the initial configurable aspects of the component.

### Related issues
Relates to RET-1295

### Backwards compatibility
Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)