refactor: remove BottomSheetLegacy and replace with BottomSheet (#6437)
### Description

As the title - the BottomSheetLegacy background was not so easy to
adjust with colors alone when Kayla and I were playing with dark and
light mode colors, and instead of trying to make it work I think it's
better to delete it.

### Test plan

No meaningful changes, I checked all the new bottom sheets manually on
my iOS simulator.

### Related issues

- Related to RET-1293

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)