refactor(colors): replace backgroundInverse with scrim and qr colors (#6442)
### Description

This PR replaces `backgroundInverse` with other colors, since the
concept is not really aligned with how Kayla is using the colors. Now we
have:
1. `backgroundScrim` for bottom sheets and underlays
2. `qrTabBar` primary and secondary - these colors are specific to the
QR tab segmented control and Kayla suggested that we are able to control
them separately.

### Test plan

n/a

### Related issues

- Related to RET-1293

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)