refactor(colors): remove shadow color (#6444)
### Description

The `shadow` color isn't really used in designs by Kayla and doesn't
make sense to keep in the palette. The two places it's used can be
switched without much visual difference.

### Test plan

These are the two places that were changed, and proof they look okay
(although the secure send content really needs to be updated)

![Simulator Screenshot - iPhone 15 Pro - 2025-01-23 at 18 13
35](https://github.com/user-attachments/assets/e5466fec-309e-41bb-bb5d-10575df8fc66)
![Simulator Screenshot - iPhone 15 Pro - 2025-01-23 at 18 12
55](https://github.com/user-attachments/assets/e6dcbc16-2124-46c2-af13-78258c7f211d)

### Related issues

- Related to RET-1293

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)