refactor(colors): replace contentInverse (#6445)
### Description

Replacing `contentInverse` since this isn't a useful semantic variable
for design. Replaced with `contentTertiary` for content on colored
backgrounds specifically, and re-evaluated the places that used
contentInverse (for the filter chips and enter amount percentage chips,
they were replaced with button colors since they should be thought of as
buttons. for the swap arrows, use backgroundPrimary because it is the
contrast that is needed)

### Test plan

No visual changes for valora

### Related issues

- Related to RET-1293

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)