refactor(colors): primary and secondary borders (#6446)
### Description

This PR does a pretty straight replace between `border` and
`borderPrimary` (since that is primary use case for Valora) but there
are a couple of places where `borderSecondary` is used (this is mainly
to facilitate beefy / dark mode designs). Kayla and I talked for some
time about how borders are used and this convention of using it with
background colors is more aligned with how she has been designing.

### Test plan

No visual changes for Valora

### Related issues

- Related to RET-1293

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)