refactor(colors): update component color usages to work for dark mode (#6447)
### Description

The changes from this PR are from doing a final pass through the app to
ensure alignment with the valora and [beefy
designs](https://www.figma.com/design/xJ4CFXy7dL0oijQD3kSqrE/Mobile-Stack---Beefy?node-id=2626-136369&t=dVwNj3eC8dlq9bNc-0).
The tricky part of finding a color palette that works for both is that
there are some design choices that are different for light vs dark
themes, in particular dark themes don't usually make use of borders and
instead use contrast backgrounds. Note that the beefy designs are not up
to date where text inputs are using white borders - those should not
have borders, and use the pale blue background instead.

As a result, this PR is a hodge podge collection of changes after my
working session with Kayla:
1. a `textInputBackground` is introduced and applied to text inputs, the
TextInput components (and consumers that add styles to it) were updated
to use this variable + the `borderSecondary` border
2. a bunch of borders were changed to `borderSecondary` - truth be told
the border rules are a little weak and Kayla + I had a really hard time
defining the border color semantics. There is probably a need to define
more border color variables, but there are already so many variables
that we decided to leave it like this for now. `borderSecondary` is more
subtle, and `borderPrimary` is more bold. The choice for which one to
use affects mainly dark mode, it'd probably be helpful to document these
rules somewhere.
3. in an effort to remove the use of borders, the fee containers on all
the enter amount screens have had the borders.
4. the dapps / earn pool / pool info cards now have secondary background
(the only visual change for valora)

### Test plan

Walkthroughs of the apps are attached to the task
[here](https://linear.app/valora/issue/RET-1293/update-color-palette-to-be-semantic#comment-667f3369)

### Related issues

- Fixes RET-1293

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)