fix: recovery phrase input colors (#6451)
### Description

As the title - I broke the colors for this component in light mode. I
took some liberty to simplify the styles and remove the shadow since
we're moving away from shadows. The colors don't exactly match the
designs but i think the semantics of the color variables makes sense
here and double checked with Kayla that the result is okay.

### Test plan

![Simulator Screenshot - iPhone 15 Pro - 2025-01-24 at 17 53
42](https://github.com/user-attachments/assets/a5e629b5-d1a8-43ea-a044-de536097a002)
![Simulator Screenshot - iPhone 15 Pro - 2025-01-24 at 17 53
32](https://github.com/user-attachments/assets/d193fe3e-f834-437c-8c8f-321de5b101ea)

### Related issues

- Fixes RET-1293

#[issue number here]

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)