refactor(colors): replace enum with object and type (#6456)
### Description

This PR changes the `colors` from enum to object. The usage is still the
same, but the object cannot be referenced as a type in the same way as
an enum can. This change facilitates being able to overwrite the default
colors with custom colors via config, plus there are benefits of moving
away from enums (e.g.
[here](https://www.totaltypescript.com/erasable-syntax-only)).

There are a lot of changes but they are repetitve, the key changes are:
1. in the colors file, Colors is now an object
2. in the colors file, also export a `ColorValue` type
3. in all files that were using `Colors` or `colors` as a type, now use
`ColorValue`

### Test plan

n/a

### Related issues

- Related to RET-1280

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)