ci: auto publish to NPM (#23)
### Description

Publish automatically to NPM using
[semantic-release](https://github.com/semantic-release/semantic-release)

For now all commits on main will be published as alpha pre-releases
(e.g. `1.0.0-alpha.X`, `X` increasing with each commit)
We can adapt / change this once we're ready to make a stable release.

Note: semantic-release doesn't support monorepo out-of-the-box
But I really like it:
- it doesn't require a git commit to bump the version in `package.json`,
which is annoying to bypass when branch protection is setup.
- it just works with conventional commits
- we use it in all our other repos that publish automatically to NPM

After reviewing many similar tools, I landed on using a community
maintained extension which supports monorepo:

[@anolilab/multi-semantic-release](https://github.com/anolilab/semantic-release/blob/main/packages/multi-semantic-release/README.md)
It's been maintained for some time now and used by a bunch of other
projects.

See also:
https://dev.to/antongolub/the-chronicles-of-semantic-release-and-monorepos-5cfc

Part of RET-1287